move memory.h to utilities
