+++ engine/include/genebits/engine/ecs/storage.h
@@ -3,8 +3,8 @@
 
 #include "genebits/engine/containers/type_map.h"
 #include "genebits/engine/containers/vector.h"
-#include "genebits/engine/os/memory.h"
 #include "genebits/engine/utilities/erased_ptr.h"
+#include "genebits/engine/utilities/memory.h"
 #include "genebits/engine/utilities/type_info.h"
 
 namespace genebits::engine
